add exported flag to declaration when default
Add ReflectionFlag.Export flag to a node's flag variable when the export
converter is run.

This fixes a scenario where `export default` is used in a separate
statement from the variable creation, such as:

```
const foo = 'bar';
export default foo;
```

In the above example, the export converter cannot find the symbolmapping
for the `foo` declaration because the createDeclaration code returns
null when --excludeNotExported flag is set.