Fix max-line-length tslint config
The max-line-length rule was not being enforced because of an invalid configuration value. In order to supply configuration information other than "on" or "off", the rule must be an array value. While this was silently failing in older versions of tslint, the upgrade to tslint 5.9.0 causes this malformed value to fail with an exception.

After fixing the configuration of this rule, many lines of the existing codebase were in violation. Rather than reformat a bunch of files, I bumped the maximum line length from 120 to 170, and reformatted a handful of outlier lines that were even longer than this.