Export all internal models for use outside TypeDoc
I'm currently implementing a documentation renderer on top of TypeDoc
and not having access to all internal models from outside of TypeDoc
makes for some nasty type hacks when traversing reflections. While the
corresponding properties do exist on the reflections, there's no way to
access them in a type safe manner due to lack of access to the models.