Prune non-exported declarations rather than not create them.
The earlier code would skip the creation of a declaration when the
flag excludeNotExported was turned on. This could cause problems
because it is not always possible at the time of the declaration
creation to know whether the declaration will ultimately be
exported. Consider:

const foo =  1;
export { foo }

At the time of processing the first line, it is not known that foo
will be exported.

The new code lets all declarations be created and later, if
excludeNotExported is turned on, it prunes the declarations that
were not exported.