Add Export flag.
The Export flag is used to indicate that a symbol is directly under
the effect of an `export` keyword in the TS source or directly subject
to an export declaration.

This flag is different from the Exported flag, which indicates that a
symbol is externally either directly *or* indirectly.

Class members provide a good example of the distinction between the
two flags. They never get the Export flag because they cannot be
individuall exported. However, they have the Exported flag true if
their containing class is exported.