Allow installing from git URL
The `prepare` script was introduced in npm 5 to distinguish between build steps and other prepublish tasks (which should be run from the `prepublishOnly` script.

npm no longer runs the `prepublish` script when installing from git repositories, but it will install dev deps and run `prepare` if it's present. Both are run before publishing.