Resolve most unit test issues
Also updates launch.json and tasks.json to be more helpful in VSCode and enables downlevelIteration.

In order to fix some of the unit tests, the target was updated to es2015 - since the lib already references es2015 libraries and we are only testing on node 6+, I believe this is acceptable.