Reset the state of the comment plugin on conversion start. (#1006)
* Revert "rebuild specs"

This reverts most of commit 3716ba7918bce8165e864d40ebf365cded3e3ee1.

* Reset the hidden field at conversion start.

In the test suite, a single app is reused for all the converter tests. If the
plugins do not appropriately reset their state at the start of a conversion,
then one test may affect later tests.

* Fix remaining specs that need fixing.

These were not affected by the revert in the previous commit.