Support linking to module.d.ts (#1117)
* Support linking to module

See my question on https://stackoverflow.com/q/58429798/1076463

Creating a link from `firstmodule.d.ts` file to the module `secondmodule.d.ts` is impossible:

- The name of the second module that you need to use is `"secondmodule.d"`, e.g. `{@link "secondmodule.d"}` (including the quotes)
- However, the name gets split on the `.`, so in the `ProjectReflection#findReflectionByName`, you end up comparing `"secondmodule.d"` with just `d` due to the `const names: string[] = Array.isArray(arg) ? arg : arg.split('.');` line

This proposed change strips the `.d` from the name, allowing you to use `{@link "secondmodule"}`.

Ideally I would also strip the quotes, but that causes a lot of test failures.
My proposed change seems to pass all existing tests, although it affects the json file that gets generated. This sounds like a backwards incompatible change, but no idea how strict the project is on this.

I also haven't quite figured out how to add a test for this :-(

If this PR is not accepted because you don't want to change the outputted json, can you tell me (or answer on SO) how to create that link.

* Fixing linting problems

Fixing linting problems

* refactor: use const instead of let where possible

* fix: also correct typo in test assertion

* Don't change the name of the module, but use a custom split method instead

* Fixed linter errors

* The splitUnquotedString function can now deal with strings where only a part is quoted

* Added test for splitUnquotedString function

* The special split function is needed in multiple places

* Made the split method more robust

* Ignore tarball, which is the result of running npm pack

* Typo

Typo

* Ignore tarball, which is the result of running npm pack

* Support linking to module

See my question on https://stackoverflow.com/q/58429798/1076463

Creating a link from `firstmodule.d.ts` file to the module `secondmodule.d.ts` is impossible:

- The name of the second module that you need to use is `"secondmodule.d"`, e.g. `{@link "secondmodule.d"}` (including the quotes)
- However, the name gets split on the `.`, so in the `ProjectReflection#findReflectionByName`, you end up comparing `"secondmodule.d"` with just `d` due to the `const names: string[] = Array.isArray(arg) ? arg : arg.split('.');` line

This proposed change strips the `.d` from the name, allowing you to use `{@link "secondmodule"}`.

Ideally I would also strip the quotes, but that causes a lot of test failures.
My proposed change seems to pass all existing tests, although it affects the json file that gets generated. This sounds like a backwards incompatible change, but no idea how strict the project is on this.

I also haven't quite figured out how to add a test for this :-(

If this PR is not accepted because you don't want to change the outputted json, can you tell me (or answer on SO) how to create that link.

* Fixing linting problems

Fixing linting problems

* Don't change the name of the module, but use a custom split method instead

* Fixed linter errors

* The splitUnquotedString function can now deal with strings where only a part is quoted

* Added test for splitUnquotedString function

* The special split function is needed in multiple places

* Made the split method more robust