Distinguish between block and other tags when parsing comments
When parsing the contents of comments, TypeDoc treats all tags as block
tags. That is, if it encounters a tag, it parses whatever follows it as
tag contents, until it reaches the end of the comment or finds another
tag.

According to [TSDoc
standard](https://tsdoc.org/pages/spec/tag_kinds) this is correct
behaviour for block tags but not for modifier or inline tags. This PR
modifies the behaviour by ensuring that only block tags are parsed this
way and that for modifier or inline tags parsing of tag content stops
when we encounter an empty line.

As a side effect, when some JSDoc tags were used at the beginning of
the comment (e.g. `@class`), the entire docstring was often removed.
This is because TypeDoc also attempts to remove redundant JSDoc tags
from the comments. Erroneous assignment of the entire comment as the
tag content meant that everything was deleted. Following this fix, only
the actual tag content (e.g. the class name for `@class` tag) will be
removed.

Closes <some-issue-number>.