fix(fs): glob will traverse node_modules if asked to
If the `pattern` passed to `glob()` contains `node_modules`, then `node_modules` will be traversed.

This is _naive_ in that it does not care about the _context_ of `node_modules` in the pattern; it could be a negation, or whatever (e.g., some passes `!**/node_modules/**`).  However, `minimatch` should still do its thing, so hopefully this won't be too bad.