Move from marked to markdown-it
Marked has served TypeDoc well for many years, but with Marked 7.0.4, it
is possible for Marked plugins to prevent parsing from being done in a sync
method, which either means that TypeDoc has to render all text to markdown
outside of the normal rendering flow, or refactor that rendering flow to support
async work within random JSX element construction.

markdown-it is also a fully featured markdown parser which has avoided async
code, making it a better fit for TypeDoc. It also makes it ridiculously easy to add
additional rendering capability via plugins.