Revert cache added in #2386
The router implementation used by TypeDoc now indicates that in a 22.5 second run
rendering all of `@types/node`, 65ms is spent getting URLs. The extra complexity
in attempting to cache the result of this frankly isn't worth it.

HTML rendering took 16.8 seconds, of which 13.6 seconds was just doing syntax
highlighting. Shiki is the most obvious place to spend time investigating performance
improvements at this point.