Make plugin peerDependency filtering more strict
Yes, this means that even plugins which use even a *tiny* bit of TypeDoc's
API will require a re-release when breaking TypeDoc versions are
released in order to continue showing up in the current version. That's
completely intended; it should be up to plugin maintainers to verify support,
not a new user who tries to use it based on a soft recommendation from TypeDoc's
plugin page. Users wanting to try to use a plugin which doesn't claim support
can still do so, either with "legacy" peer deps, or by using overrides.